backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic
transformers
torch
pyttsx3
requests
python-dotenv